+++ asset_server/src/asset_sourcer/shader_sourcer.rs
@@ -1,7 +1,8 @@
 use asset_common::{shader::Shader, AssetRef};
 
 use crate::{
-    asset::AssetDependency, file_change::FileTimestamp, source_files::SourceFileRef, AssetInserter,
+    asset::AssetDependency, asset_server::AssetInserter, file_change::FileTimestamp,
+    source_files::SourceFileRef,
 };
 
 use super::{Asset, AssetSourcer, CreateAssetInfo};
